chore(rover): moves utils into utils dir
simplifies directory structure by moving
client, env, git, stringify, and telemetry
into the utils directory
