+++ src/command/graph/check.rs
@@ -4,9 +4,9 @@ use structopt::StructOpt;
 
 use rover_client::query::graph::check;
 
-use crate::client::StudioClientConfig;
 use crate::command::RoverStdout;
-use crate::git::GitContext;
+use crate::utils::client::StudioClientConfig;
+use crate::utils::git::GitContext;
 use crate::utils::loaders::load_schema_from_flag;
 use crate::utils::parsers::{parse_graph_ref, parse_schema_source, GraphRef, SchemaSource};
 use crate::Result;
