+++ src/command/subgraph/check.rs
@@ -5,9 +5,9 @@ use structopt::StructOpt;
 use crate::Result;
 use rover_client::query::subgraph::check;
 
-use crate::client::StudioClientConfig;
 use crate::command::RoverStdout;
-use crate::git::GitContext;
+use crate::utils::client::StudioClientConfig;
+use crate::utils::git::GitContext;
 use crate::utils::loaders::load_schema_from_flag;
 use crate::utils::parsers::{parse_graph_ref, parse_schema_source, GraphRef, SchemaSource};
