+++ src/command/subgraph/fetch.rs
@@ -4,8 +4,8 @@ use structopt::StructOpt;
 
 use rover_client::query::subgraph::fetch;
 
-use crate::client::StudioClientConfig;
 use crate::command::RoverStdout;
+use crate::utils::client::StudioClientConfig;
 use crate::utils::parsers::{parse_graph_ref, GraphRef};
 use crate::Result;
