+++ src/utils/mod.rs
@@ -1,2 +1,7 @@
+pub mod client;
+pub mod env;
+pub mod git;
 pub mod loaders;
 pub mod parsers;
+pub mod stringify;
+pub mod telemetry;
